CORE-620 Removed "extended" package. Moved all functions to "util".
