+++ rspace/src/main/scala/coop/rchain/rspace/util/package.scala
@@ -43,4 +43,20 @@ package object util {
     val (l1, l2) = xs.span(x => !p(x))
     l1 ++ (l2 drop 1)
   }
+
+  /**
+    * Extracts a continuation from
+    */
+  def getK[A, K](t: Option[(K, A)]): K =
+    t.map(_._1).get
+
+  /** Runs a continuation with the accompanying data
+    */
+  def runK[T](t: Option[(Function1[T, Unit], T)]): Unit =
+    t.foreach { case (k, data) => k(data) }
+
+  /** Runs a list of continuations with the accompanying data
+    */
+  def runKs[T](t: Seq[Option[(Function1[T, Unit], T)]]): Unit =
+    t.foreach { case Some((k, data)) => k(data); case None => () }
 }
