+++ rspace/src/test/scala/coop/rchain/rspace/ConcurrencyTests.scala
@@ -2,7 +2,7 @@ package coop.rchain.rspace
 
 import coop.rchain.rspace.examples.AddressBookExample._
 import coop.rchain.rspace.examples.AddressBookExample.implicits._
-import coop.rchain.rspace.extended._
+import coop.rchain.rspace.util._
 import monix.eval.Task
 import monix.execution.Scheduler.Implicits.global
 import scala.concurrent.Await
